+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import {defineConfig} from 'tsup';
 
 export default defineConfig({
-  entry: ['src/cli.ts'],
+  entry: ['src/cli.ts', 'src/config.ts'],
   format: ['esm'],
   platform: 'node',
   target: 'node24',